# Enable CUDA graph train on torchvision models