--- a/torchbenchmark/util/framework/vision/args.py
+++ b/torchbenchmark/util/framework/vision/args.py
@@ -9,7 +9,7 @@ def parse_args(model: BenchmarkModel, extra_args: List[str]) -> argparse.Namespa
     parser.add_argument("--eval-fp16", action='store_false', help="enable eval fp16")
     parser.add_argument("--fx2trt", action='store_true', help="enable fx2trt")
     parser.add_argument("--flops", action='store_true', help="enable flops counting")
-    parser.add_argument("--train_cudagraph", action='store_true', help="enable CUDA Graph for train")
+    parser.add_argument("--cudagraph", action='store_true', help="enable CUDA Graph. Currently only implemented for train.")
     args = parser.parse_args(extra_args)
     args.device = model.device
     args.jit = model.jit
@@ -34,7 +34,7 @@ def apply_args(model: BenchmarkModel, args: argparse.Namespace):
         assert not args.jit, "fx2trt with JIT is not available."
         model.eval_model = enable_fx2trt(args.eval_bs, args.eval_fp16, model.eval_model, model.eval_example_inputs)
     # apply cuda graph for train
-    if args.train_cudagraph:
+    if args.cudagraph:
         enable_cudagraph(model, model.example_inputs)
 
 def enable_cudagraph(model: BenchmarkModel, example_inputs: Tuple[torch.tensor]):